refactor(runtime): extract shared types
