+++ ui/src/lib/runtime/index.ts
@@ -1,0 +1,6 @@
+export type {
+  RuntimeAudioObjectDescriptor,
+  RuntimeAudioObjectService
+} from './types/audio-adapter';
+
+export type { RuntimeObjectDescriptor } from './types/runtime-object';
